// frontend/src/app/eventoscaza/models/evento-caza.ts
export interface EventoCaza {
  cazaId: string;
  fechaEvento: string;
  tipoEvent: string;
  lugar: string;
  ciudad: string;
  horaEvento: string;
  comentariosEvento: string;
  agenda: string;
  personasInvitadaHref: string;
  tipoCaza: string;
  precinto: string;
  trofeo: string;
  cotoCaza: string;
  zonasMarcadasHref: string;
  perrosCazaHref: string;
  agendaHref: string;
}